haccp_app/requirements.txt
streamlit>=1.30
pandas>=2.2
reportlab>=4.0
psycopg2-binary>=2.9
